+++ src/Adapters/AtlasQuery.php
@@ -41,7 +41,7 @@ class AtlasQuery implements AdapterInterface
     /**
      * @inheritDoc
      */
-    public function findAll(string $table, ?array $where = null, ?string $orderBy = null): array
+    public function findAll(string $table, ?array $where = null, ?string $orderBy = null, ?int $limit = null, ?int $offset = null): array
     {
         $select = Select::new($this->pdo);
 
@@ -51,6 +51,12 @@ class AtlasQuery implements AdapterInterface
 
         if ($orderBy)
             $query->orderBy($orderBy);
+
+        if ($limit)
+            $query->limit($limit);
+
+        if ($offset)
+            $query->offset($offset);
 
         return $query->fetchAll();
     }